feat: Create basic categories setting file #395
