+++ server/package.json
@@ -16,7 +16,6 @@
     "@types/morgan": "^1.9.4",
     "@types/passport-google-oauth20": "^2.0.11",
     "@types/winston": "^2.4.4",
-    "bcrypt": "^5.1.0",
     "connect-redis": "^6.1.3",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
@@ -46,7 +45,6 @@
     "@babel/core": "^7.22.1",
     "@babel/preset-env": "^7.22.2",
     "@babel/preset-typescript": "^7.21.5",
-    "@types/bcrypt": "^5.0.0",
     "@types/connect-redis": "^0.0.20",
     "@types/cookie-parser": "^1.4.3",
     "@types/cors": "^2.8.13",
